move module rest to hyle-modules
