+++ crates/hyle-modules/Cargo.toml
@@ -10,7 +10,6 @@ rust-version = { workspace = true }
 
 [dependencies]
 sdk = { workspace = true, default-features = false, features = ["full-model"] }
-hyle-crypto = { workspace = true }
 client-sdk = { workspace = true, features = ["rest", "indexer"] }
 hyle-net = { workspace = true }
 hyle-verifiers = { workspace = true }
@@ -23,12 +22,10 @@ tracing = "0.1"
 strum_macros = "0.27.1"
 serde_with = "3.12.0"
 config = { version = "0.15.9", default-features = false, features = ["toml"] }
-sha3 = "0.10.8"
 assertables = { version = "9.5.2" }
 anymap = { version = "0.12.1" }
 syn = { version = "2.0.101" }
 quote = { version = "1.0.39" }
-opentelemetry = { version = "0.28" }
 rand = { version = "0.8.5" }
 futures = { version = "0.3.31" }
 tracing-subscriber = { version = "0.3", default-features = false, features = [
@@ -38,13 +35,22 @@ tracing-subscriber = { version = "0.3", default-features = false, features = [
 ] }
 paste = { version = "1.0.15" }
 hex = "0.4.3"
-
 bytes = { version = "1.10.0" }
+
 tokio = { version = "1.44.2", features = ["full", "tracing"] }
 tokio-util = { version = "0.7.14" }
+
 axum = { version = "0.8.4", features = ["ws"] }
+axum-otel-metrics = { version = "0.10.0" }
+
 utoipa = { version = "5.3.1" }
 utoipa-axum = { version = "0.2.0" }
+utoipa-swagger-ui = { version = "9.0.0", features = ["axum"] }
+
+tower-http = { version = "0.6.2", features = ["trace", "cors"] }
+prometheus = { version = "0.13.4" }
+opentelemetry = { version = "0.28" }
+
 test-log = { version = "0.2.17", features = [
   "color",
   "trace",
